Create command to grab last 100 messages
The command grabs the last 100 messages in a channel and then uploads a
json file containing the message objects in the same channel.

Right now I'm just poking around to see how I want this command to work.
I haven't figured out exactly the representation for archiving channel
information that isn't verbose info that doesn't mean much to a user.