Allow command to grab all messages in a channel
Change the name of the command from scrape to archive. Implement
grabbing all the message objects in a channel instead of just the last
100.

The bot doesn't save the json or upload it to the channel anymore but
it is just commented out. There's no need for it right now since there's
a lot of useless, for my needs, information included and it'll just be
cluttered with that. Wait until you extract the stuff you want to save
it and upload it to the channel.