Send metadata information about the guild, channel
When the archive command is called, the bot will get all the messages in
the channel and also get some select guild and channel information and
store it into an object.

The information is not sent still. Right now only some guild and channel
info is retreived. No message info, besides the snowflakes, is gathered
and placed into the object that holds all the information we are going
to send.