Extract and store message information
The messages gathered earlier are now extracted for information and are
stored in the object to be delivered.

This basically just handles text only messages, reactions, threads,
images, embeds, etc. are not handled so the information isn't captured
at all. Also, the actual sending of the file is still commented out
since it's a bit annoying to test with and there has to be a way to
ignore previous resulting archive .json files generated from the archive
command.