Add usage and description for archive command
Check to see if the argument supplied to the archive command is valid.
Print the usage and description to the console when appropriate. Create
two arguments, help and text, that the user needs to choose from when
calling this command.

The messages print to the console for now. I don't want to spam the bot
testing channel too much.