Reduce generated .json by restructuring data
Reduce the size of the generated json file by around 50% by renaming the
fields that store the message data. Also reduce redundancy of
information by adding a participants field to the channel_data so that
the pfp url and the user snowflake is only stored once.

The size of the file is going to be an issue so keep this in mind. Think
about what can be stored in one place or how the file can be
restructured.