Update usage and description of archive command
Update the usage and description of the archive command to better
reflect what it should end up resembling.

There are 4 arguments that do specific things and should ignore other
arguments. I think it would be important to have metadata and
participants be things we can extract without messages so they should
each be one of these four arguments. Help is obvious, should ignore
everything. The complete argument should also not necessitate any
further arguments since the other arguments specify how much to capture.
I think it's important to have a messages-only flag to specify that we
want to get only the message information. For channels with many
messages it would be difficult to export and open such a large file only
to get rid of guild and channel information to make processing it
easier. So this option seems necessary. The whole-messages argument is a
shortcut for all the arguments: text, reactions, stickers, attachments.