Add command args validation and execution dispatch
Validate the arguments that were passed to the archive command and send
messages to the channel when something goes wrong. Add a dispatching
switch statement skeleton that will choose what to do based on the
arguments passed. Send messages to the channel if not implemented yet.
Fix the usage and description. Add a new recognized-arguments array that
will be used later to implement a way to detect unrecognized arguments.

I'm planning right now to check for unrecognized arguments in the main
magi file.