Re-enable the text argument to work again
Get rid of the archive_channel function and replace it with a new
function that will later figure out what to archive based on the
arguments that were passed to the command. Rename a few things.

The goal here is to make one function reusable by the complete, text,
and whole-messages arguments. This won't be difficult to do since
complete and whole-messages are almost aliases for text with all
capture arguments chosen. Also, I'm not entirely sure whether the
participants argument is reasonable. I'll think about it a bit more
though. I think there might be a way.