Insert comments with alias code for data gathering
Add code that is commented out in the "complete" and "whole-message"
argument cases. This code will pretty much serve as an alias for text
with all the options enabled. With the special case where
"whole-message" was used with the "messages-only" argument. That
argument will be added in if it was used.

Comment out this code and remove the unimplemented messages when all the
"text" arguments: "reactions", "stickers", and "attachments" work
correctly. They should be properly implemented after it is done.