Implement the messages-only argument
Restructure the get_data function to handle the messages-only argument.
Also add args as an argument to the get_message_data function so that it
knows what to extract from messages.

The only function that needs to be worked on now is the get_message_data
function. Everything else will fall into place.