Uncomment to implement complete, whole-messages
Uncomment the code to implement the complete and whole-messages
arguments. Also create the timestamp to capture the time when the
command was called since the messages are fetched from newest to oldest.

This looks pretty solid right now but there is definitely room for
improvement such as adding in stickere capture and reducing file size.