Selectively capture thread information
Add an option to be able to decide to choose to capture thread messages.
Update the usage, description, and alias commands to take this into
account. Capture whether or not a message spawns a thread. Capture the
id of the thread if the message spawns a thread or belongs to a thread.
Update participant info if a member participated in a thread but not in
the channel. Also fix an issue where a message capture would show that
it is a reply when that is not the case.

This is a, in my opinion, really messy implementation I have so far. I
think the best thing to do right now is to add in support for messages
with more than 1 attachment and to capture the guild icon. After this I
think that's all the functionality I currently want / can test for since
I don't have discord nitro. Refactoring will be the next major step and
then the command will be pretty much complete for this version of
discord (minus stickers).