Capture sticker information (untested)
Capture sticker information if the user selects the stickers argument.
Remove message sent about stickers not being implemented in the message
data gathering function. Capture information similar to how information
was captured for attachments. Include fields to capture creator of the
sticker as well.

This is not tested in the sense that there are no stickers on my test
guild.