Split message aggregation and data extraction
Make the extraction of desired message data into its own function so
that we can extract from any message collection and not directly from a
channel. Comment out any code related to capturing thread information.

The reason for this is so that we have more control over when the data
extraction from messages happens. And we can change when we want to
extract if need be. Capturing thread data is a bit complicated right now
and I'd rather not deal with it until later when I can focus on it as
its own thing.