Change function names and disable sending files
Change function names to better reflect their function in the future
when functions will generate files first and then send once they are all
generated.

The next step is to generate the metadata file. After that make changes
to the generateChannelFiles function that will adapt to the changes
made here.