Add file generation promises for channel messages
Add an array to keep track of pages generated from channel messages and
return a promise indicating that all files were generated. Do this in
the function responsible for generating the message files.

This was a great lesson in fulfill vs resolve. This function should
still be usable by the part in our program to handle getting messages
from threads. So what should happen is that there should be a parameter
passed that lets us switch between modes: one for the whole channel
itself and one mode that behaves in a certain way if it's in a thread.
This way, we can ensure we only generate pages in the context of the
channel as a whole and ignore doing this when we are in a channel. This
is why there are two different things this function can return.