Implement page generation when json would be 8MB+
If the result of the json file would be over 8MB then generate the page
containing the message data. Empty the collection and continue command
execution. Generate the last page if necessary and also generate
participant information.

There's a lot of room for improvement but for now focus on implementing
the gzipping and tarring of json files and sending. Also implement the
participants argument when possible.