Export command structure check to main magi.js
Export the isValidCommand function from the archive command file to be
used in magi.js.

Each function has it's own structure and the current check does not
support arguments that aren't predetermined. So this way of doing things
should allow each command to check its own structure. Although maybe
there are better ways to do this.