Update information in tokenizer-info
Add information about how a command name must be followed by a space character.

The decision was made to require one space after the command name if there
are arguments given. There is no need to have anything besides a space as
whitespace from the perspective of a user entering a command.